Move embedded plugins to the Gradle Plugins classloader
by moving them to the :provider-plugins project
and adding 2 SPI interfaces in order to share code with :provider
