+++ provider/src/test/kotlin/org/gradle/kotlin/dsl/accessors/GenerateProjectSchemaTest.kt
@@ -2,6 +2,7 @@ package org.gradle.kotlin.dsl.accessors
 
 import org.gradle.kotlin.dsl.provider.spi.ProjectSchema
 import org.gradle.kotlin.dsl.provider.spi.ProjectSchemaEntry
+import org.gradle.kotlin.dsl.provider.spi.loadMultiProjectSchemaFrom
 
 import org.gradle.kotlin.dsl.fixtures.AbstractIntegrationTest
